+++ tests/api_resources/test_positions.py
@@ -17,7 +17,7 @@ base_url = os.environ.get("TEST_API_BASE_URL", "http://127.0.0.1:4010")
 class TestPositions:
     parametrize = pytest.mark.parametrize("client", [False, True], indirect=True, ids=["loose", "strict"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     def test_method_list(self, client: ParsecAPI) -> None:
         position = client.positions.list(
@@ -25,16 +25,17 @@ class TestPositions:
         )
         assert_matches_type(PositionListResponse, position, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     def test_method_list_with_all_params(self, client: ParsecAPI) -> None:
         position = client.positions.list(
             exchange="exchange",
             market_id="market_id",
+            x_exchange_credentials="X-Exchange-Credentials",
         )
         assert_matches_type(PositionListResponse, position, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     def test_raw_response_list(self, client: ParsecAPI) -> None:
         response = client.positions.with_raw_response.list(
@@ -46,7 +47,7 @@ class TestPositions:
         position = response.parse()
         assert_matches_type(PositionListResponse, position, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     def test_streaming_response_list(self, client: ParsecAPI) -> None:
         with client.positions.with_streaming_response.list(
@@ -66,7 +67,7 @@ class TestAsyncPositions:
         "async_client", [False, True, {"http_client": "aiohttp"}], indirect=True, ids=["loose", "strict", "aiohttp"]
     )
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     async def test_method_list(self, async_client: AsyncParsecAPI) -> None:
         position = await async_client.positions.list(
@@ -74,16 +75,17 @@ class TestAsyncPositions:
         )
         assert_matches_type(PositionListResponse, position, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     async def test_method_list_with_all_params(self, async_client: AsyncParsecAPI) -> None:
         position = await async_client.positions.list(
             exchange="exchange",
             market_id="market_id",
+            x_exchange_credentials="X-Exchange-Credentials",
         )
         assert_matches_type(PositionListResponse, position, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     async def test_raw_response_list(self, async_client: AsyncParsecAPI) -> None:
         response = await async_client.positions.with_raw_response.list(
@@ -95,7 +97,7 @@ class TestAsyncPositions:
         position = await response.parse()
         assert_matches_type(PositionListResponse, position, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     async def test_streaming_response_list(self, async_client: AsyncParsecAPI) -> None:
         async with async_client.positions.with_streaming_response.list(