+++ tests/api_resources/test_trades.py
@@ -17,7 +17,7 @@ base_url = os.environ.get("TEST_API_BASE_URL", "http://127.0.0.1:4010")
 class TestTrades:
     parametrize = pytest.mark.parametrize("client", [False, True], indirect=True, ids=["loose", "strict"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     def test_method_list(self, client: ParsecAPI) -> None:
         trade = client.trades.list(
@@ -25,11 +25,12 @@ class TestTrades:
         )
         assert_matches_type(TradeListResponse, trade, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     def test_method_list_with_all_params(self, client: ParsecAPI) -> None:
         trade = client.trades.list(
             parsec_id="parsec_id",
+            cursor="cursor",
             end_ts=0,
             limit=1,
             outcome="outcome",
@@ -37,7 +38,7 @@ class TestTrades:
         )
         assert_matches_type(TradeListResponse, trade, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     def test_raw_response_list(self, client: ParsecAPI) -> None:
         response = client.trades.with_raw_response.list(
@@ -49,7 +50,7 @@ class TestTrades:
         trade = response.parse()
         assert_matches_type(TradeListResponse, trade, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     def test_streaming_response_list(self, client: ParsecAPI) -> None:
         with client.trades.with_streaming_response.list(
@@ -69,7 +70,7 @@ class TestAsyncTrades:
         "async_client", [False, True, {"http_client": "aiohttp"}], indirect=True, ids=["loose", "strict", "aiohttp"]
     )
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     async def test_method_list(self, async_client: AsyncParsecAPI) -> None:
         trade = await async_client.trades.list(
@@ -77,11 +78,12 @@ class TestAsyncTrades:
         )
         assert_matches_type(TradeListResponse, trade, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     async def test_method_list_with_all_params(self, async_client: AsyncParsecAPI) -> None:
         trade = await async_client.trades.list(
             parsec_id="parsec_id",
+            cursor="cursor",
             end_ts=0,
             limit=1,
             outcome="outcome",
@@ -89,7 +91,7 @@ class TestAsyncTrades:
         )
         assert_matches_type(TradeListResponse, trade, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     async def test_raw_response_list(self, async_client: AsyncParsecAPI) -> None:
         response = await async_client.trades.with_raw_response.list(
@@ -101,7 +103,7 @@ class TestAsyncTrades:
         trade = await response.parse()
         assert_matches_type(TradeListResponse, trade, path=["response"])
 
-    @pytest.mark.skip(reason="Prism tests are disabled")
+    @pytest.mark.skip(reason="Mock server tests are disabled")
     @parametrize
     async def test_streaming_response_list(self, async_client: AsyncParsecAPI) -> None:
         async with async_client.trades.with_streaming_response.list(